Vu+ multiboot environment (OV)
VuplusKexec.py: update embedded skin

- Added self.footnote to specify the RootInit action.
- The common Title widget has been removed, in this way we do not see repeated text in the RootInit action and title, separate text is written in the footnote of the action that is executed + detailed description.
- Update FlashImage help text.

MultiBoot.py: define UUID - UUIDnum to getCurrentImage

FlashImage.py: remove num slot is defined in MultiBoot UUIDnum

VuplusKexec.py: add exception Vu+ Zero 4k

- This machine automatically starts flash if it detects the vuplus folder with its image files inside on a device.
- So when we enable MultiBoot and press GREEN it reboots, if the user has connected a device with this folder it will automatically flash again, then eMMC slots will not be restored.
- A footnote is added with the text of the guideline to be followed by the user if he encounters this situation.

FlashImage.py: sort list of kexec slots - add description

FlashImage - MultiBoot: Improve environment USB Slots kexec

- Avoid Device problems (lost) for the initialization of Slots. Use QUIT REBOOT to add new USB slots.
- Added USB eMMC SlotType prefix on items (thanks to ViX)
- Added try sentence for key device for ensure mount exist, same with download images.
- Variable NumberSlots 4 slots for non-Kexec systems, there will be 3 eMMc slots for kexec and flash, we do not include slot0.

+++ lib/python/Screens/VuplusKexec.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from os.path import join
+from Components.Harddisk import harddiskmanager
 from Components.ActionMap import ActionMap
 from Components.Console import Console
 from Components.Label import Label
@@ -19,16 +21,12 @@ STARTUP_3 = "kernel=/linuxrootfs3/zImage root=/dev/%s rootsubdir=linuxrootfs3" %
 class VuplusKexec(Screen):
 
 	skin = """
-	<screen name="VuplusKexec" position="center,center" size="750,700" flags="wfNoBorder" backgroundColor="transparent">
-		<eLabel name="b" position="0,0" size="750,700" backgroundColor="#00ffffff" zPosition="-2" />
-		<eLabel name="a" position="1,1" size="748,698" backgroundColor="#00000000" zPosition="-1" />
-		<widget source="Title" render="Label" position="center,14" foregroundColor="#00ffffff" size="e-10%,35" horizontalAlignment="left" verticalAlignment="center" font="Regular; 28" backgroundColor="#00000000" />
-		<eLabel name="line" position="1,60" size="748,1" backgroundColor="#00ffffff" zPosition="1" />
-		<eLabel name="line2" position="1,250" size="748,4" backgroundColor="#00ffffff" zPosition="1" />
-		<widget source="description" render="Label" position="2,80" size="730,30" horizontalAlignment="center" font="Regular; 22" backgroundColor="#00000000" foregroundColor="#00ffffff" />
-		<!-- widget source="key_red" render="Label" position="30,200" size="150,30" noWrap="1" zPosition="1" verticalAlignment="center" font="Regular; 20" horizontalAlignment="left" backgroundColor="#00000000" foregroundColor="#00ffffff" / -->
-		<widget source="key_green" render="Label" position="200,200" size="150,30" noWrap="1" zPosition="1" verticalAlignment="center" font="Regular; 20" horizontalAlignment="left" backgroundColor="#00000000" foregroundColor="#00ffffff" />
-		<ePixmap pixmap="skin_default/buttons/green.png" position="200,200" size="40,40" alphaTest="blend" />
+	<screen name="VuplusKexec" position="center,180" size="980,235" resolution="1280,720">
+		<widget source="footnote" render="Label" position="0,10" size="960,50" foregroundColor="yellow" horizontalAlignment="center" font="Regular;20" />
+		<widget source="description" render="Label" position="0,70" size="960,500" horizontalAlignment="center" font="Regular;22"  />
+		<widget source="key_green" render="Label" objectTypes="key_green,StaticText" position="20,e-170" size="180,40" backgroundColor="key_green" conditional="key_green" font="Regular;18" foregroundColor="key_text" horizontalAlignment="center" verticalAlignment="center">
+			<convert type="ConditionalShowHide" />
+		</widget>
 	</screen>
 	"""
 
@@ -37,6 +35,7 @@ class VuplusKexec(Screen):
 		self.title = _("Vu+ MultiBoot Manager")
 		self["description"] = StaticText(_("Press GREEN button to enable MultiBoot\n\nWill reboot within 10 seconds,\nunless you have eMMC slots to restore.\nRestoring eMMC slots can take from 1 -> 5 minutes per slot."))
 		# self["key_red"] = StaticText(_("Cancel"))
+		self["footnote"] = StaticText()
 		self["key_green"] = StaticText(_("Init Vu+ MultiBoot"))
 		self["actions"] = ActionMap(["SetupActions"],
 		{
@@ -45,31 +44,39 @@ class VuplusKexec(Screen):
 			"cancel": self.close,
 			"menu": self.close,
 		}, -1)
+		partitions = sorted(harddiskmanager.getMountedPartitions(), key=lambda partitions: partitions.device or "")
+		for partition in partitions:
+			self.folderVuplus = join(partition.mountpoint, "vuplus")
+			if BoxInfo.getItem("model") == "vuzero4k" and pathExists(self.folderVuplus):
+				self["footnote"].text = _("Delete or rename folder \"vuplus\" in path '%s' before initializing.") % partition.mountpoint
 
 	def RootInit(self):
-		self["actions"].setEnabled(False)  # This function takes time so disable the ActionMap to avoid responding to multiple button presses
-		if fileExists("/usr/bin/kernel_auto.bin") and fileExists("/usr/bin/STARTUP.cpio.gz"):
-			self.title = _("Vu+ MultiBoot Initialisation - will reboot after 10 seconds.")
-			self["description"].text = _("Vu+ MultiBoot Initialisation in progress\n\nWill reboot after restoring any eMMC slots\nThis can take from 1 -> 5 minutes per slot.")
-			with open("/STARTUP", 'w') as f:
-				f.write(STARTUP)
-			with open("/STARTUP_RECOVERY", 'w') as f:
-				f.write(STARTUP_RECOVERY)
-			with open("/STARTUP_1", 'w') as f:
-				f.write(STARTUP_1)
-			with open("/STARTUP_2", 'w') as f:
-				f.write(STARTUP_2)
-			with open("/STARTUP_3", 'w') as f:
-				f.write(STARTUP_3)
-			print("[VuplusKexec][RootInit] Kernel Root", BoxInfo.getItem("mtdkernel"), "   ", BoxInfo.getItem("mtdrootfs"))
-			cmdlist = []
-			cmdlist.append("dd if=/dev/%s of=/zImage" % BoxInfo.getItem("mtdkernel"))						# backup old kernel
-			cmdlist.append("dd if=/usr/bin/kernel_auto.bin of=/dev/%s" % BoxInfo.getItem("mtdkernel"))  # create new kernel
-			cmdlist.append("mv /usr/bin/STARTUP.cpio.gz /STARTUP.cpio.gz")						# copy userroot routine
-			Console().eBatch(cmdlist, self.RootInitEnd, debug=True)
-		else:
-			self.session.open(MessageBox, _("VuplusKexec: Create Vu+ Multiboot environment - Unable to complete, Vu+ Multiboot files missing."), MessageBox.TYPE_INFO, timeout=30)
+		if BoxInfo.getItem("model") == "vuzero4k" and pathExists(self.folderVuplus):
 			self.close()
+		else:
+			self["actions"].setEnabled(False)  # This function takes time so disable the ActionMap to avoid responding to multiple button presses
+			if fileExists("/usr/bin/kernel_auto.bin") and fileExists("/usr/bin/STARTUP.cpio.gz"):
+				self["footnote"].text = _("Vu+ MultiBoot Initialisation - will reboot after 10 seconds.")
+				self["description"].text = _("Vu+ MultiBoot Initialisation in progress\n\nWill reboot after restoring any eMMC slots\nThis can take from 1 -> 5 minutes per slot.")
+				with open("/STARTUP", 'w') as f:
+					f.write(STARTUP)
+				with open("/STARTUP_RECOVERY", 'w') as f:
+					f.write(STARTUP_RECOVERY)
+				with open("/STARTUP_1", 'w') as f:
+					f.write(STARTUP_1)
+				with open("/STARTUP_2", 'w') as f:
+					f.write(STARTUP_2)
+				with open("/STARTUP_3", 'w') as f:
+					f.write(STARTUP_3)
+				print("[VuplusKexec][RootInit] Kernel Root", BoxInfo.getItem("mtdkernel"), "   ", BoxInfo.getItem("mtdrootfs"))
+				cmdlist = []
+				cmdlist.append("dd if=/dev/%s of=/zImage" % BoxInfo.getItem("mtdkernel"))						# backup old kernel
+				cmdlist.append("dd if=/usr/bin/kernel_auto.bin of=/dev/%s" % BoxInfo.getItem("mtdkernel"))  # create new kernel
+				cmdlist.append("mv /usr/bin/STARTUP.cpio.gz /STARTUP.cpio.gz")						# copy userroot routine
+				Console().eBatch(cmdlist, self.RootInitEnd, debug=True)
+			else:
+				self.session.open(MessageBox, _("VuplusKexec: Create Vu+ Multiboot environment - Unable to complete, Vu+ Multiboot files missing."), MessageBox.TYPE_INFO, timeout=30)
+				self.close()
 
 	def RootInitEnd(self, *args, **kwargs):
 		print("[VuplusKexec][RootInitEnd] rebooting")
